move IrcSubscriber to Subscriber namespace
used only internally; safe to move
